Add files via upload
•	Briefly summarize the requirements and goals of the app you developed. What user needs was this app designed to address?
The app I created was designed to help users track their daily weight and alert them once they reached their goal weight. This app included a database with tables for daily weight entries that could be modified, user logins, passwords, and their goal weight. The app would also send SMS notifications when certain conditions were met- such as when the user reached their goal weight.  The app would also let users create a new account if they did not already have one. 
•	What screens and features were necessary to support user needs and produce a user-centered UI for the app? How did your UI designs keep users in mind? Why were your designs successful?
The app had a login screen for user authentication, a screen that would allow users to log daily weight, a screen that displayed date, and weight, and allowed users to adjust as well as a user profile manager to alert them when they reached their goal weight. 
My design’s goal was to be a User-Centered UI design that would be simple and easy to use.  I want it to be a straightforward tracker that would let them log and track their daily weight in an intuitive manner.  It was important to focus on how to implement the grid view for the daily weight entries along with the corresponding dates as well as provide SMS notifications.
•	How did you approach the process of coding your app? What techniques or strategies did you use? How could those be applied in the future?
For this app, I utilized Android Studio, Java, and SQLite for database management. When it comes to strategy, I realized it’s important to figure out design flow, and UX design, and understand user goals before starting any project.  Code should be readable, easy to understand and well commented for others.
•	How did you test to ensure your code was functional? Why is this process important and what did it reveal?
To ensure the code was functional, I performed rigorous testing with the Android Emulators to ensure all my features worked as they should.  I tested various user scenarios with adding and modifying weight, as well as SMS permissions.  I quickly realized the importance of testing to resolve any bugs/issues and to ensure a smooth user experience.
•	Considering the full app design and development process, from initial planning to finalization, where did you have to innovate to overcome a challenge?
I have never used Android Studio or SQLite, have some experience with Java, so it was a learning experience how to figure out how to plan out and go about the coding for this project. 
•	In what specific component of your mobile app were you particularly successful in demonstrating your knowledge, skills, and experience?
Some skills I feel I gained were how to better use Android Studio, the start of a database for database management, and the ability to try and create a user-centered UI with an emphasis on usability and simplicity.   I also realized the importance of constant testing to ensure all is running as it should I plan on utilizing much of what I’ve used in this course for future studies.